introduction of an abstract base class
