+++ src/main/java/ch/ethz/idsc/retina/dev/rimo/RimoRateController.java
@@ -6,7 +6,7 @@ import ch.ethz.idsc.tensor.Scalar;
 import ch.ethz.idsc.tensor.qty.Quantity;
 
 /** two instance of this class are used for left and right rear motors
- * @see RimoRateControllerWrap
+ * @see RimoRateControllerDuo
  * 
  * Kp with unit "ARMS*rad^-1*s"
  * Ki with unit "ARMS*rad^-1" */
