+++ src/main/java/ch/ethz/idsc/retina/dev/zhkart/joy/RimoRateJoystickModule.java
@@ -3,18 +3,15 @@ package ch.ethz.idsc.retina.dev.zhkart.joy;
 
 import java.util.Optional;
 
-import ch.ethz.idsc.owly.car.math.DifferentialSpeed;
 import ch.ethz.idsc.retina.dev.joystick.GokartJoystickInterface;
 import ch.ethz.idsc.retina.dev.rimo.RimoConfig;
 import ch.ethz.idsc.retina.dev.rimo.RimoPutEvent;
-import ch.ethz.idsc.retina.dev.rimo.RimoRateControllerWrap;
+import ch.ethz.idsc.retina.dev.rimo.RimoRateControllerDuo;
 import ch.ethz.idsc.retina.dev.rimo.RimoSocket;
 import ch.ethz.idsc.retina.dev.steer.SteerColumnInterface;
 import ch.ethz.idsc.retina.dev.steer.SteerConfig;
 import ch.ethz.idsc.retina.dev.steer.SteerSocket;
-import ch.ethz.idsc.retina.gui.gokart.top.ChassisGeometry;
 import ch.ethz.idsc.tensor.Scalar;
-import ch.ethz.idsc.tensor.Tensor;
 
 /** DO NOT USE THIS IMPLEMENTATION.
  * 
@@ -32,7 +29,7 @@ import ch.ethz.idsc.tensor.Tensor;
 // @Deprecated
 /* package */ class RimoRateJoystickModule extends JoystickModule<RimoPutEvent> {
   private final SteerColumnInterface steerColumnInterface = SteerSocket.INSTANCE.getSteerColumnTracker();
-  /* package */ final RimoRateControllerWrap rimoRateControllerWrap = new RimoRateControllerWrap();
+  /* package */ final RimoRateControllerDuo rimoRateControllerWrap = new RimoRateControllerDuo();
 
   @Override // from AbstractModule
   void protected_first() {
@@ -59,11 +56,11 @@ import ch.ethz.idsc.tensor.Tensor;
       GokartJoystickInterface joystick) {
     if (steerColumnInterface.isSteerColumnCalibrated()) {
       Scalar speed = RimoConfig.GLOBAL.rateLimit.multiply(joystick.getAheadAverage());
-      DifferentialSpeed differentialSpeed = ChassisGeometry.GLOBAL.getDifferentialSpeed();
+      // DifferentialSpeed differentialSpeed = ChassisGeometry.GLOBAL.getDifferentialSpeed();
       Scalar theta = SteerConfig.GLOBAL.getAngleFromSCE(steerColumnInterface);
-      Tensor pair = differentialSpeed.pair(speed, theta);
-      Tensor bias = joystick.getAheadPair_Unit().multiply(RimoConfig.GLOBAL.rateLimit);
-      return rimoRateControllerWrap.iterate(pair.add(bias));
+      // Tensor pair = differentialSpeed.pair(speed, theta);
+      // Tensor bias = joystick.getAheadPair_Unit().multiply(RimoConfig.GLOBAL.rateLimit);
+      return rimoRateControllerWrap.iterate(speed, theta);
     }
     return Optional.empty();
   }
