+++ src/main/java/ch/ethz/idsc/retina/dev/zhkart/pure/PurePursuitRimo.java
@@ -3,25 +3,20 @@ package ch.ethz.idsc.retina.dev.zhkart.pure;
 
 import java.util.Optional;
 
-import ch.ethz.idsc.owly.car.math.DifferentialSpeed;
 import ch.ethz.idsc.retina.dev.rimo.RimoPutEvent;
 import ch.ethz.idsc.retina.dev.rimo.RimoPutProvider;
+import ch.ethz.idsc.retina.dev.rimo.RimoRateControllerDuo;
 import ch.ethz.idsc.retina.dev.rimo.RimoRateControllerWrap;
-import ch.ethz.idsc.retina.dev.rimo.RimoRateControllerWrap_2;
 import ch.ethz.idsc.retina.dev.rimo.RimoSocket;
 import ch.ethz.idsc.retina.dev.steer.SteerColumnInterface;
 import ch.ethz.idsc.retina.dev.steer.SteerConfig;
 import ch.ethz.idsc.retina.dev.steer.SteerSocket;
 import ch.ethz.idsc.retina.dev.zhkart.ProviderRank;
-import ch.ethz.idsc.retina.gui.gokart.top.ChassisGeometry;
-import ch.ethz.idsc.tensor.Scalar;
-import ch.ethz.idsc.tensor.Tensor;
 
 class PurePursuitRimo extends PurePursuitBase implements RimoPutProvider {
   private final SteerColumnInterface steerColumnInterface = SteerSocket.INSTANCE.getSteerColumnTracker();
-  /* package */ final RimoRateControllerWrap rimoRateControllerWrap = new RimoRateControllerWrap();
-  final RimoRateControllerWrap_2 rimoRateControllerWrap_2 = new RimoRateControllerWrap_2();
-  private final boolean useDiffSpeed = false;
+  /* package */ final RimoRateControllerWrap rimoRateControllerWrap = //
+      new RimoRateControllerDuo();
 
   @Override // from StartAndStoppable
   public void start() {
@@ -44,18 +39,10 @@ class PurePursuitRimo extends PurePursuitBase implements RimoPutProvider {
   }
 
   /* package */ Optional<RimoPutEvent> control(SteerColumnInterface steerColumnInterface) {
-    if (steerColumnInterface.isSteerColumnCalibrated()) {
-      Scalar speed = PursuitConfig.GLOBAL.rateFollower;
-      if (useDiffSpeed == true) {
-        // TODO either kill this completely or make a way to change control mode through the interface
-        DifferentialSpeed differentialSpeed = ChassisGeometry.GLOBAL.getDifferentialSpeed();
-        Scalar theta = SteerConfig.GLOBAL.getAngleFromSCE(steerColumnInterface);
-        Tensor pair = differentialSpeed.pair(speed, theta);
-        return rimoRateControllerWrap.iterate(pair);
-      } else {
-        return rimoRateControllerWrap_2.iterate(speed);
-      }
-    }
+    if (steerColumnInterface.isSteerColumnCalibrated())
+      return rimoRateControllerWrap.iterate( //
+          PursuitConfig.GLOBAL.rateFollower, // average target velocity
+          SteerConfig.GLOBAL.getAngleFromSCE(steerColumnInterface)); // steering angle
     return Optional.empty();
   }
