+++ src/test/java/ch/ethz/idsc/retina/gui/gokart/RimoMetronomeModule.java
@@ -6,14 +6,13 @@ import java.util.Optional;
 import ch.ethz.idsc.owl.data.Stopwatch;
 import ch.ethz.idsc.retina.dev.rimo.RimoPutEvent;
 import ch.ethz.idsc.retina.dev.rimo.RimoPutProvider;
-import ch.ethz.idsc.retina.dev.rimo.RimoRateControllerWrap;
+import ch.ethz.idsc.retina.dev.rimo.RimoRateControllerDuo;
 import ch.ethz.idsc.retina.dev.rimo.RimoSocket;
 import ch.ethz.idsc.retina.dev.zhkart.ProviderRank;
 import ch.ethz.idsc.retina.sys.AbstractModule;
 import ch.ethz.idsc.tensor.RealScalar;
 import ch.ethz.idsc.tensor.Scalar;
 import ch.ethz.idsc.tensor.Scalars;
-import ch.ethz.idsc.tensor.Tensors;
 import ch.ethz.idsc.tensor.qty.Quantity;
 import ch.ethz.idsc.tensor.sca.Mod;
 
@@ -23,7 +22,7 @@ class RimoMetronomeModule extends AbstractModule implements RimoPutProvider {
   public static final Scalar HALF_PERIOD = RealScalar.of(2);
   // ---
   private final Stopwatch stopwatch = Stopwatch.started();
-  private final RimoRateControllerWrap rimoRateControllerWrap = new RimoRateControllerWrap();
+  private final RimoRateControllerDuo rimoRateControllerWrap = new RimoRateControllerDuo();
   private final Scalar testPulseLo = Quantity.of(0, "rad*s^-1");
   private final Scalar testPulseHi = Quantity.of(20, "rad*s^-1");
 
@@ -50,8 +49,6 @@ class RimoMetronomeModule extends AbstractModule implements RimoPutProvider {
     Scalar remaind = Mod.function(HALF_PERIOD.multiply(RealScalar.of(2))).apply(seconds);
     boolean isPassive = Scalars.lessThan(remaind, HALF_PERIOD);
     // ---
-    return rimoRateControllerWrap.iterate(Tensors.of( //
-        isPassive ? testPulseLo : testPulseHi, //
-        isPassive ? testPulseLo : testPulseHi));
+    return rimoRateControllerWrap.iterate(isPassive ? testPulseLo : testPulseHi, RealScalar.ZERO);
   }
 }
